AUC scores of models

Dataset auc_scores contains a subset of binary classification models results from Kretowicz and Biecek (2020).

model - name of model followed by a number indicating setting of hyperparameters

split - indicator of split

auc - area under the ROC curve

data(auc_scores)

Format

a data frame with 1600 rows and 3 columns.